# openapi.json
{
  "components": {
    "schemas": {
      "CounterCoverage": {
        "description": "How much of the fabric the window actually saw.\n\n`unsupported` is fabric-level and per counter, because a counter no port can\nreport is a property of the hardware rather than of any one port -- and\nsaying it once keeps it off every port row. It is what lets the UI grey\nout a checkbox instead of showing a column of honest, useless zeros.",
        "properties": {
          "ports_expected": {
            "title": "Ports Expected",
            "type": "integer"
          },
          "ports_measured": {
            "title": "Ports Measured",
            "type": "integer"
          },
          "ports_no_data": {
            "title": "Ports No Data",
            "type": "integer"
          },
          "ports_partial": {
            "title": "Ports Partial",
            "type": "integer"
          },
          "ports_reset": {
            "title": "Ports Reset",
            "type": "integer"
          },
          "sweeps": {
            "title": "Sweeps",
            "type": "integer"
          },
          "unsupported": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "symbol_error",
                "link_error_recovery",
                "link_downed",
                "rcv_errors",
                "rcv_remote_phys_errors",
                "rcv_switch_relay_errors",
                "xmit_discards",
                "xmit_constraint_errors",
                "rcv_constraint_errors",
                "local_link_integrity",
                "excessive_buffer_overrun",
                "vl15_dropped",
                "xmit_wait",
                "qp1_dropped"
              ]
            },
            "title": "Unsupported",
            "type": "object"
          }
        },
        "required": [
          "ports_expected",
          "ports_measured",
          "ports_no_data",
          "ports_reset",
          "ports_partial",
          "sweeps"
        ],
        "title": "CounterCoverage",
        "type": "object"
      },
...
      "ErrorDeltas": {
        "properties": {
          "counters": {
            "items": {
              "enum": [
                "symbol_error",
                "link_error_recovery",
                "link_downed",
                "rcv_errors",
                "rcv_remote_phys_errors",
                "rcv_switch_relay_errors",
                "xmit_discards",
                "xmit_constraint_errors",
                "rcv_constraint_errors",
                "local_link_integrity",
                "excessive_buffer_overrun",
                "vl15_dropped",
                "xmit_wait",
                "qp1_dropped"
              ],
              "type": "string"
            },
            "title": "Counters",
            "type": "array"
          },
          "coverage": {
            "$ref": "#/components/schemas/CounterCoverage"
          },
          "fabric": {
            "title": "Fabric",
            "type": "string"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/PortErrorDelta"
            },
            "title": "Ports",
            "type": "array"
          },
          "window": {
            "$ref": "#/components/schemas/CounterWindow"
          }
        },
        "required": [
          "fabric",
          "window",
          "coverage",
          "counters",
          "ports"
        ],
        "title": "ErrorDeltas",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinkDetail": {
        "properties": {
          "a": {
            "$ref": "#/components/schemas/LinkEnd"
          },
          "b": {
            "$ref": "#/components/schemas/LinkEnd"
          },
          "health": {
            "enum": [
              "ok",
              "degraded",
              "down",
              "unknown"
            ],
            "title": "Health",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "items": {
              "type": "string"
            },
            "title": "Reason",
            "type": "array"
          },
          "resolved": {
            "$ref": "#/components/schemas/Resolved"
          }
        },
        "required": [
          "resolved",
          "id",
          "health",
          "a",
          "b"
        ],
        "title": "LinkDetail",
        "type": "object"
      },
...
      "NodeDetail": {
        "properties": {
          "desc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desc"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "first_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "fw_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fw Version"
          },
          "guid": {
            "title": "Guid",
            "type": "string"
          },
          "health": {
            "enum": [
              "ok",
              "degraded",
              "down",
              "unknown"
            ],
            "title": "Health",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "num_ports": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Ports"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/PortDetail"
            },
            "title": "Ports",
            "type": "array"
          },
          "resolved": {
            "$ref": "#/components/schemas/Resolved"
          },
          "sm_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sm State"
          },
          "system_image_guid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Image Guid"
          },
          "type": {
            "enum": [
              "ca",
              "switch",
              "router"
            ],
            "title": "Type",
            "type": "string"
          },
          "vendor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor"
          },
          "vendor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          }
        },
        "required": [
          "resolved",
          "guid",
          "label",
          "type",
          "health"
        ],
        "title": "NodeDetail",
        "type": "object"
      },
...
      "PortErrorDelta": {
        "description": "One port's accumulation over the window.\n\nSparse in two directions: `d` carries only counters that moved, and a port\nwith nothing to say is absent entirely. That deliberately inverts the\nstorage decision -- the schema stores dense because columnar compression\nrewards it, while the wire is sparse because the client already holds the\nport universe and a healthy fabric is nearly all zeros.\n\nAbsence therefore means zero, which is only safe because every case where it\nwould instead mean *unknown* raises a flag below.\n\nEvery field below `port` keeps a default and the route serialises with\n`exclude_defaults`, so a clean row is `{\"node\", \"port\", \"d\"}`. The\nenvelope's fields stay required so the same switch cannot elide a zero that\nmeans something.",
        "properties": {
          "d": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "symbol_error",
                "link_error_recovery",
                "link_downed",
                "rcv_errors",
                "rcv_remote_phys_errors",
                "rcv_switch_relay_errors",
                "xmit_discards",
                "xmit_constraint_errors",
                "rcv_constraint_errors",
                "local_link_integrity",
                "excessive_buffer_overrun",
                "vl15_dropped",
                "xmit_wait",
                "qp1_dropped"
              ]
            },
            "title": "D",
            "type": "object"
          },
          "no_data": {
            "default": false,
            "title": "No Data",
            "type": "boolean"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "partial": {
            "default": false,
            "title": "Partial",
            "type": "boolean"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "reset": {
            "default": false,
            "title": "Reset",
            "type": "boolean"
          },
          "unsupported": {
            "items": {
              "enum": [
                "symbol_error",
                "link_error_recovery",
                "link_downed",
                "rcv_errors",
                "rcv_remote_phys_errors",
                "rcv_switch_relay_errors",
                "xmit_discards",
                "xmit_constraint_errors",
                "rcv_constraint_errors",
                "local_link_integrity",
                "excessive_buffer_overrun",
                "vl15_dropped",
                "xmit_wait",
                "qp1_dropped"
              ],
              "type": "string"
            },
            "title": "Unsupported",
            "type": "array"
          }
        },
        "required": [
          "node",
          "port"
        ],
        "title": "PortErrorDelta",
        "type": "object"
      },
...
      "TrafficRates": {
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "cadence_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cadence S"
          },
          "coverage": {
            "$ref": "#/components/schemas/TrafficCoverage"
          },
          "fabric": {
            "title": "Fabric",
            "type": "string"
          },
          "first_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First At"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/PortTraffic"
            },
            "title": "Ports",
            "type": "array"
          },
          "requested_window_s": {
            "title": "Requested Window S",
            "type": "number"
          },
          "span_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span S"
          }
        },
        "required": [
          "fabric",
          "as_of",
          "requested_window_s",
          "coverage",
          "ports"
        ],
        "title": "TrafficRates",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Read-only view of the InfiniBand fabric, at any point in time.",
    "title": "ibmon read API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/fabrics/{fabric}/counters/errors": {
      "get": {
        "description": "Per-port error deltas over `(since, until]`, sparse.\n\nA port is present only if a counter moved or a flag is raised; everything\nabsent was measured, is zero, and is trusted. `coverage` is what makes that\nsafe to read -- check it before believing an empty `ports`.",
        "operationId": "errors_api_v1_fabrics__fabric__counters_errors_get",
        "parameters": [
          {
            "in": "path",
            "name": "fabric",
            "required": true,
            "schema": {
              "title": "Fabric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "description": "Seconds to look back before `since` for the baseline reading. 0 disables the asof baseline.",
            "in": "query",
            "name": "lookback",
            "required": false,
            "schema": {
              "default": 900,
              "description": "Seconds to look back before `since` for the baseline reading. 0 disables the asof baseline.",
              "maximum": 86400,
              "minimum": 0,
              "title": "Lookback",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDeltas"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Errors",
        "tags": [
          "fabric"
        ]
      }
    },
    "/api/v1/fabrics/{fabric}/counters/traffic": {
      "get": {
        "description": "Per-port transmit and receive rates over a trailing window, sparse.\n\nA rate, not a delta: `at` names the instant to read *at*, and `window` only\nsays how long an interval to average over. Ports below the idle floor are\nomitted; ports that could not be measured are present and flagged, because\nan unpolled port and an idle one must not render alike.",
        "operationId": "traffic_api_v1_fabrics__fabric__counters_traffic_get",
        "parameters": [
          {
            "in": "path",
            "name": "fabric",
            "required": true,
            "schema": {
              "title": "Fabric",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          },
          {
            "description": "Trailing window in seconds. Defaults to three times the fabric's observed traffic cadence.",
            "in": "query",
            "name": "window",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 3600,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Trailing window in seconds. Defaults to three times the fabric's observed traffic cadence.",
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrafficRates"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Traffic",
        "tags": [
          "fabric"
        ]
      }
    },
...
    "/api/v1/fabrics/{fabric}/links/{link_id}": {
      "get": {
        "description": "Both ends expanded in place, so selecting a link is one request.\n\nThe id is the canonical GUID-ordered form; the database's own port_id\nordering never appears on the wire, so this parses without consulting it.",
        "operationId": "link_api_v1_fabrics__fabric__links__link_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "fabric",
            "required": true,
            "schema": {
              "title": "Fabric",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "link_id",
            "required": true,
            "schema": {
              "title": "Link Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          },
          {
            "description": "Seconds of counter history to summarise on each port. Omit for no counters. Pass the same window the overlay is using, or the card and the graph will disagree.",
            "in": "query",
            "name": "counters_window",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 86400,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Seconds of counter history to summarise on each port. Omit for no counters. Pass the same window the overlay is using, or the card and the graph will disagree.",
              "title": "Counters Window"
            }
          },
          {
            "description": "Include per-port throughput. The window is server-sized from the fabric's observed traffic cadence, matching the overlay.",
            "in": "query",
            "name": "with_traffic",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include per-port throughput. The window is server-sized from the fabric's observed traffic cadence, matching the overlay.",
              "title": "With Traffic",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Link",
        "tags": [
          "fabric"
        ]
      }
    },
    "/api/v1/fabrics/{fabric}/nodes/{guid}": {
      "get": {
        "operationId": "node_api_v1_fabrics__fabric__nodes__guid__get",
        "parameters": [
          {
            "in": "path",
            "name": "fabric",
            "required": true,
            "schema": {
              "title": "Fabric",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "guid",
            "required": true,
            "schema": {
              "title": "Guid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          },
          {
            "description": "Seconds of counter history to summarise on each port. Omit for no counters. Pass the same window the overlay is using, or the card and the graph will disagree.",
            "in": "query",
            "name": "counters_window",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 86400,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Seconds of counter history to summarise on each port. Omit for no counters. Pass the same window the overlay is using, or the card and the graph will disagree.",
              "title": "Counters Window"
            }
          },
          {
            "description": "Include per-port throughput. The window is server-sized from the fabric's observed traffic cadence, matching the overlay.",
            "in": "query",
            "name": "with_traffic",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include per-port throughput. The window is server-sized from the fabric's observed traffic cadence, matching the overlay.",
              "title": "With Traffic",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Node",
        "tags": [
          "fabric"
        ]
      }
    },
...
  }
}